dns: add some additional error checks.
If the upper 2 bits of a label/pointer value are 10, treat that as an
error.

If a name is longer than 255 characters, treat that as an error.

This prevents some long loops with malformed packets, as found by Hardik
Shah.

(backported from commit 92d636a906d450f9bd344ee312cfa9c88c3d2bd6)