Build system: Add WITH_COVERAGE to compile netdissect and tcpdump with code coverage
Build with --coverage and use the coveralls tool to submit
code coverage reports to https://coveralls.io/github/the-tcpdump-group/tcpdump

Support for coverage builds is added to both automake and cmake, but
coverage relies on running tests, and running tests is currently broken
with cmake.  So the added job is CMAKE=no but can be changed to
CMAKE=yes when running tests with cmake is fixed.