Fix a build problem on the Linux Travis CI buildbot.
The buildbot is failing in autotools builds - but not in CMake builds;
go figure.  Perhaps they're passing different flags to the C compiler,
and, in the autotools builds, the flag causes the compiler to define
whatever it takes to get /usr/include/features.h to define
_DEFAULT_SOURCE as 1, so that our defining it as nothing collides with
that.

(Dear UN*X community: please come up with a platform-independent,
reliable way of saying "expose every single API you have, namespace
pollution be damned.  kthxbye.)