EAP: back out the two trailing-comma changes.
Hopefully this provokes one more build on the buildbots, with a matching
print-eap.c and tests/eapon1-v-nosmb.out, so that "make check" succeeds.

Step 2 is to check the new versions back in, in one commit, in the hopes
that "make check" succeeds.

It appears that, on all the buildbots, print-esp.c is up to date but
tests/eapon1-v-nosmb.out isn't; I have *NO* idea how that happened, but
it again reminds me that I hate software.