Avoid calling setuid() with -Zroot
Fixes an issue under Linux user namespaces, where having a uid of 0
is not enough for setuid(0) to succeed making tcpdump would exit.
There was no way to avoid this when compiled -DWITH_USER='"tcpdump"'.

This patch achieves the behaviour documented in source by avoiding
the setuid() when -Zroot is supplied.

FIxes this scenario on Linux:

  $ unshare -Urn
  # ip link add br0 type bridge
  # ip link set br0 up
  # tcpdump -ibr0 -lvn
  tcpdump: Couldn't change to 'tcpdump' uid=137 gid=146: Operation not permitted
  # tcpdump -Zroot -ibr0 -lvn
  tcpdump: Couldn't change to 'root' uid=0 gid=0: Operation not permitted