instrument functions: Enhance the output
If entering in a function, print also the calling function name with
file name and line number. There may be a small shift in the line number.

In some cases, with Clang 11, the file number is unknown (printed '??')
or the line number is unknown (printed '?'). In this case, use GCC.

To print nothing, like with no instrumentation:
$ make instrument_off

As before, the following commands are available:
To configure the printing of only the global functions names:
$ make instrument_global

To go back to print all the functions names:
$ make instrument_all

The library libbfd is used, therefore the binutils-dev package is required.