juniper: make the cookie-processing code a little clearer.
In the loop searching through the cookie table, just search for a
matching entry and, if we find one, save the entry and break out of the
loop.

If we found a matching entry, process the cookie using that entry.

That way, the cookie processing isn't buried in a loop.

(cherry picked from commit 57be4eb7ed4974965910d0ec179540b59bfe7a19)