Makefile.in: don't remove configure and config.h.in in make distclean.
In the GNU makefile standards, "make distclean" removes everything
that's not part of the source distribution; this does *not* include the
configure script or the config.h.in file, as those are in the release
tarball, so that building from a source tarball doesn't require having
autoconf.  This allows somebody building from the source tarball to do a
"make distclean" and then reconfigure.

Do, however, remove any release tarball in make distclean, by having
distclean depend on clean.