autoconf: Retire AC_LBL_C_INLINE. [skip appveyor]
In 2002 this macro implemented a workaround for HP C compiler because
the latter did not work with the implementation of AC_C_INLINE in
Autoconf 2.13 (see commit b1263c6).  Since then the required kind of
inline very likely became available in every supported compiler of every
supported OS, but just in case there is still an exception, use
AC_C_INLINE from Autoconf 2.69 (or later) and trust it to work right.
In other words, if the original problem still exists, this will have to
be proven.

This squelches one more warning from Autoconf 2.71:

configure.ac:32: warning: The macro `AC_TRY_COMPILE' is obsolete.