configure: don't use egrep, use $EGREP.
POSIX does not require that a system support egrep; it requires that it
support grep -E.  In AC_LBL_LIBPCAP, use AC_PROG_EGREP to check for a
command that supports egrep-style expressions, and use $EGREP rather
than egrep in subsequent commands.

This squelches warnings on platforms where egrep prints a warning
suggesting that grep -E be used instead.