Autoconf: Remove detection of early IPv6 stacks. [skip appveyor]
The KAME/INRIA/etc. block has been around since commit c9d84d1 in 1999,
when it was common for IPv6 stacks to exist and develop separately from
operating systems.  During the next 10 or so years IPv6 support in
various operating systems became the norm and the stack detector became
obsolete.  Remove it and continue to use libc IPv6 implementation.